betting-chatbot/web/backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
pyjwt==2.6.0
gunicorn==20.1.0 